fixup ANDROID: of_get_flat_dt_prop

+++ drivers/of/fdt.c
@@ -911,7 +911,7 @@ static const char *config_cmdline = "";
 int __init early_init_dt_scan_chosen(unsigned long node, const char *uname,
 				     int depth, void *data)
 {
-	unsigned long l = 0;
+	int l = 0;
 	char *p = NULL;
 	char *cmdline = data;
 
@@ -942,7 +942,7 @@ int __init early_init_dt_scan_chosen(unsigned long node, const char *uname,
 			strncpy(cmdline + cmdline_len, p, copy_len);
 			cmdline[cmdline_len + copy_len] = '\0';
 		} else {
-			strlcpy(cmdline, p, min((int)l, COMMAND_LINE_SIZE));
+			strlcpy(cmdline, p, min(l, COMMAND_LINE_SIZE));
 		}
 	}
